fix:bit-shift error

+++ src/ext4_impls/extents.rs
@@ -410,7 +410,7 @@ impl Ext4 {
             newex.first_block = to + 1;
             newex.block_count = (ee_block + block_count as u32 - 1 - to) as u16;
             newex.start_lo = (newblock & 0xffffffff) as u32;
-            newex.start_hi = (newblock >> 32) as u16;
+            newex.start_hi = ((newblock as u64) >> 32) as u16;
 
             self.insert_extent(inode_ref, &mut newex)?;
